UPDATE: Updated changes for new minor release
Updated docs link in README.
Added function to batch_convert - however, this function is not
implemented.
Added update to img_dir_list function.

The updates in batch_convert and the img_dir_list function are intended
to address the same issue, but in different ways.

The batch_convert dir_clean_up function is intended to recursively
search each directory and remove the indexing files. This would occur
prior to the image file search. Should the parent study image directory
be large - this could significantly extend the runtime. This function
also assumes that the user has write access to the parent study image
directory (which is not always the case).

The img_dir_list update would simply skip indexing files, while the
image file search is being conducted. The benefit in this case is that
only one search is performed, and that only read access is required.